複数選択とソートも
