+++ package.json
@@ -13,32 +13,32 @@
 		"test": "vitest"
 	},
 	"devDependencies": {
-		"@skeletonlabs/skeleton": "2.9.0",
-		"@skeletonlabs/tw-plugin": "0.3.1",
+		"@skeletonlabs/skeleton": "^2.9.2",
+		"@skeletonlabs/tw-plugin": "0.4.0",
 		"@sveltejs/adapter-auto": "^2.1.1",
 		"@sveltejs/kit": "^1.30.4",
-		"@testing-library/svelte": "^4.2.1",
+		"@testing-library/svelte": "^4.2.3",
 		"@types/jest": "^29.5.12",
-		"@types/node": "20.11.27",
+		"@types/node": "20.12.7",
 		"@typescript-eslint/eslint-plugin": "^6.21.0",
 		"@typescript-eslint/parser": "^6.21.0",
 		"@vite-pwa/sveltekit": "^0.4.0",
-		"autoprefixer": "10.4.18",
+		"autoprefixer": "10.4.19",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^8.10.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.38.0",
 		"jsdom": "^22.1.0",
-		"postcss": "8.4.35",
+		"postcss": "8.4.38",
 		"prettier": "^2.8.8",
 		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.2.12",
-		"svelte-check": "^3.6.7",
-		"svelte-dnd-action": "^0.9.42",
-		"tailwindcss": "3.4.1",
+		"svelte": "^4.2.15",
+		"svelte-check": "^3.6.9",
+		"svelte-dnd-action": "^0.9.43",
+		"tailwindcss": "3.4.3",
 		"tslib": "^2.6.2",
-		"typescript": "^5.4.2",
+		"typescript": "^5.4.5",
 		"vite": "^4.5.3",
-		"vite-plugin-tailwind-purgecss": "0.2.0",
+		"vite-plugin-tailwind-purgecss": "0.3.1",
 		"vitest": "^0.34.6"
 	},
 	"type": "module",
@@ -48,12 +48,11 @@
 		"@sveltejs/adapter-vercel": "^3.1.0",
 		"@types/graceful-fs": "^4.1.9",
 		"@vercel/analytics": "^1.2.2",
-		"nostr-login": "^1.0.17",
 		"nostr-tools": "^1.17.0",
 		"nosvelte": "^0.1.3",
 		"rx-nostr": "^0.9.0",
 		"svelte-i18n": "^4.0.0",
 		"unfurl.js": "^6.4.0",
-		"vite-plugin-pwa": "^0.19.3"
+		"vite-plugin-pwa": "^0.19.8"
 	}
 }
