+++ src/app.html
@@ -13,6 +13,7 @@
 
 <body data-sveltekit-preload-data="hover" data-theme="my-custom2">
 	<div style="display: contents">%sveltekit.body%</div>
+	<script src="https://www.unpkg.com/nostr-login@latest/dist/unpkg.js"></script>
 </body>
 
 </html>
