+++ src/lib/components/ListedEvent.svelte
@@ -7,7 +7,7 @@
 	import MoveIcon from '@material-design-icons/svg/round/arrow_circle_right.svg?raw';
 	import OpenIcon from '@material-design-icons/svg/round/open_in_browser.svg?raw';
 	import ShareIcon from '@material-design-icons/svg/round/chat.svg?raw';
-	import { Metadata, Nostr, NostrApp, Text, UniqueEventList } from 'nosvelte';
+
 	import { nip19, type Event as NostrEvent } from 'nostr-tools';
 	import {
 		getIdByTag,
@@ -26,7 +26,7 @@
 	import ProfileCard from './ProfileCard.svelte';
 	import Emoji from './Emoji.svelte';
 
-	import { createEventDispatcher } from 'svelte';
+	import { afterUpdate, createEventDispatcher } from 'svelte';
 	import Relay from './Relay.svelte';
 	import Other from './Other.svelte';
 	import Reference from '$lib/components/Reference.svelte';
@@ -44,6 +44,8 @@
 		type PopupSettings,
 		type ModalSettings
 	} from '@skeletonlabs/skeleton';
+	import EventandButtons from './EventandButtons.svelte';
+	import type { Nostr } from 'nosvelte';
 
 	export let DeleteNote: (e: {
 		detail: { number: number; event: any; tagArray: any };
@@ -75,13 +77,7 @@
 		}
 	};
 	console.log(bkm);
-	//let viewList: string[][];
-	//一つのタグに一種類のイベントしかないことにして日付だけ見る
-	const uniqueEvent = (eventList: NostrEvent[]): NostrEvent => {
-		//console.log(eventList);
-		eventList.sort((a, b) => b.created_at - a.created_at);
-		return eventList[0];
-	};
+
 	const privateList = async (list: NostrEvent) => {
 		if (list.content !== '') {
 			//	try {
@@ -224,6 +220,7 @@
 	$: dadClass = $isMulti === MultiMenu.Sort ? 'md:mr-0 mr-6 ' : '';
 
 	let comboboxValue: string = '';
+	let popupElement: HTMLDivElement;
 
 	const popupCombobox: PopupSettings = {
 		event: 'click',
@@ -232,12 +229,19 @@
 		closeQuery: '.listbox-item',
 		state: (test) => {
 			console.log(test);
+
 			if (!test.state) {
 				comboboxValue = '';
 			}
 		}
 	};
 
+	afterUpdate(() => {
+		if (popupElement?.style.opacity == '0') {
+			popupElement.style.top = '0';
+			popupElement.style.left = '0';
+		}
+	});
 	//-----------------------------------------------引用ポスト
 	const postNoteModalComponent: ModalComponent = {
 		ref: ModalPostNote
@@ -313,767 +317,18 @@
 							<!--なんもしない-->
 						{:else}
 							<!-- ノート | ボタン群-->
-
-							{#if tag.name[0] === 'e'}
-								<!-- {#if $searchRelays && $searchRelays.length > 0}
-					<NostrApp relays={$searchRelays}> -->
-								<Text queryKey={[id]} {id} let:text>
-									<div
-										slot="loading"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										<SearchCard
-											{filter}
-											message={`loading [${tag.name}]`}
-											isPageOwner={isOwner}
-										/>
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										slot="error"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										<SearchCard
-											{filter}
-											message={`error [${tag.name}]`}
-											isPageOwner={isOwner}
-										/>
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										slot="nodata"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										<SearchCard
-											{filter}
-											message={`not found [${tag.name}]`}
-											isPageOwner={isOwner}
-										/><button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-
-									<Metadata
-										queryKey={['metadata', text.pubkey]}
-										pubkey={text.pubkey}
-										let:metadata
-									>
-										<div
-											slot="loading"
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={text}
-												metadata={undefined}
-												{pubkey}
-											/><button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: text,
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-										<div
-											slot="error"
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={text}
-												metadata={undefined}
-												{pubkey}
-											/><button
-												class="fill-white btn-icon btn-icon-sm variant-filled-primary h-fit text-sm"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: undefined,
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-										<div
-											slot="nodata"
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={text}
-												metadata={undefined}
-												{pubkey}
-											/><button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: undefined,
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-										<div
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={text}
-												{metadata}
-												{pubkey}
-											/>
-											<button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: text,
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-									</Metadata>
-								</Text>
-							{:else if tag.name[0] === 'a'}
-								<!-- {#if $searchRelays && $searchRelays.length > 0}
-					<NostrApp relays={$searchRelays}> -->
-								<UniqueEventList
-									queryKey={tag.name}
-									filters={[filter]}
-									let:events
-								>
-									<div
-										slot="loading"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										{#if kind && kind === 30023}<!--long form content-->
-											<OGP
-												ogp={{
-													title: 'Long Form Content',
-													image: '',
-													description:
-														'open in habla' +
-														ogpDescription(parseNaddr(tag.name)),
-													favicon: 'https://habla.news/favicon.png'
-												}}
-												url={'https://habla.news/a/\n' +
-													nip19.naddrEncode(parseNaddr(tag.name))}
-											/>
-											<!---->
-										{:else if kind && kind === 34550}<!--communities-->
-											<OGP
-												ogp={{
-													title: 'Communities',
-													image: '',
-													description:
-														'open in habla\n' +
-														ogpDescription(parseNaddr(tag.name)),
-													favicon: 'https://habla.news/favicon.png'
-												}}
-												url={'https://habla.news/c/' +
-													nip19.naddrEncode(parseNaddr(tag.name))}
-											/>
-										{:else}
-											<SearchCard
-												{filter}
-												message={`loading [${tag.name}]`}
-												isPageOwner={isOwner}
-											/>
-										{/if}
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										slot="error"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										{#if kind && kind === 30023}
-											<OGP
-												ogp={{
-													title: 'Long Form Content',
-													image: '',
-													description:
-														'open in habla\n' +
-														ogpDescription(parseNaddr(tag.name)),
-													favicon: 'https://habla.news/favicon.png'
-												}}
-												url={'https://habla.news/a/' +
-													nip19.naddrEncode(parseNaddr(tag.name))}
-											/>
-											<!---->
-										{:else if kind && kind === 34550}<!--communities-->
-											<OGP
-												ogp={{
-													title: 'Communities',
-													image: '',
-													description:
-														'open in habla\n' +
-														ogpDescription(parseNaddr(tag.name)),
-													favicon: 'https://habla.news/favicon.png'
-												}}
-												url={'https://habla.news/c/' +
-													nip19.naddrEncode(parseNaddr(tag.name))}
-											/>
-										{:else}
-											<SearchCard
-												{filter}
-												message={`error [${tag.name}]`}
-												isPageOwner={isOwner}
-											/>
-										{/if}
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										slot="nodata"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										{#if kind && kind === 30023}
-											<OGP
-												ogp={{
-													title: 'Long Form Content',
-													image: '',
-													description:
-														'open in habla\n' +
-														ogpDescription(parseNaddr(tag.name)),
-													favicon: 'https://habla.news/favicon.png'
-												}}
-												url={'https://habla.news/a/' +
-													nip19.naddrEncode(parseNaddr(tag.name))}
-											/>
-											<!---->
-										{:else if kind && kind === 34550}<!--communities-->
-											<OGP
-												ogp={{
-													title: 'Communities',
-													image: '',
-													description:
-														'open in habla\n' +
-														ogpDescription(parseNaddr(tag.name)),
-													favicon: 'https://habla.news/favicon.png'
-												}}
-												url={'https://habla.news/c/' +
-													nip19.naddrEncode(parseNaddr(tag.name))}
-											/>
-										{:else}
-											<SearchCard
-												{filter}
-												message={`not found [${tag.name}]`}
-												isPageOwner={isOwner}
-											/>
-										{/if}
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-
-									<Metadata
-										queryKey={['metadata', uniqueEvent(events).pubkey]}
-										pubkey={uniqueEvent(events).pubkey}
-										let:metadata
-									>
-										<div
-											slot="loading"
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={uniqueEvent(events)}
-												metadata={undefined}
-												{pubkey}
-											/>
-											<button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: uniqueEvent(events),
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-										<div
-											slot="error"
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={uniqueEvent(events)}
-												metadata={undefined}
-												{pubkey}
-											/><button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: uniqueEvent(events),
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-										<div
-											slot="nodata"
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={uniqueEvent(events)}
-												metadata={undefined}
-												{pubkey}
-											/><button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: uniqueEvent(events),
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-										<div
-											class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-										>
-											<EventCard
-												isPageOwner={isOwner}
-												tagArray={tag.name}
-												note={uniqueEvent(events)}
-												{metadata}
-												{pubkey}
-											/>
-											<button
-												class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-												use:popup={popupCombobox}
-												on:click={() => {
-													selectedIndex = {
-														detail: {
-															number: tag.id,
-															event: uniqueEvent(events),
-															tagArray: tag.name
-														}
-													};
-												}}
-											>
-												{'Menu'}
-											</button>
-										</div>
-									</Metadata>
-								</UniqueEventList>
-							{:else if tag.name[0] === 'p'}
-								<Metadata
-									queryKey={['metadata', tag.name[1]]}
-									pubkey={tag.name[1]}
-									let:metadata
-								>
-									<div
-										slot="loading"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1 break-all"
-									>
-										<SearchCard
-											{filter}
-											message={`loading [${tag.name}]`}
-											isPageOwner={isOwner}
-										/>
-
-										<!-- loading ... {JSON.stringify(tag.name)} -->
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										slot="error"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1 break-all"
-									>
-										<SearchCard
-											{filter}
-											message={`not found [${tag.name}]`}
-											isPageOwner={isOwner}
-										/>
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										slot="nodata"
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1 break-all"
-									>
-										<SearchCard
-											{filter}
-											message={`not found [${tag.name}]`}
-											isPageOwner={isOwner}
-										/>
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											use:popup={popupCombobox}
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-										>
-											{'Menu'}
-										</button>
-									</div>
-									<div
-										class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto] gap-1"
-									>
-										<ProfileCard {metadata} tagArray={tag.name} />
-										<button
-											class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-											on:click={() => {
-												selectedIndex = {
-													detail: {
-														number: tag.id,
-														event: undefined,
-														tagArray: tag.name
-													}
-												};
-											}}
-											use:popup={popupCombobox}
-										>
-											{'Menu'}
-										</button>
-									</div>
-								</Metadata>
-							{:else if tag.name[0] === 'emoji'}
-								<!--えもじ-->
-								<div
-									class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto_auto] gap-1"
-								>
-									<Emoji tagArray={tag.name} />
-									{#if isOwner && !$isMulti}
-										<button
-											class="btn p-2 fill-surface-600 dark:fill-surface-300"
-											on:click={() => {
-												handleClick(tag.id, tag.name);
-											}}>{@html EditIcon}</button
-										>
-									{/if}
-									<button
-										class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-										use:popup={popupCombobox}
-										on:click={() => {
-											selectedIndex = {
-												detail: {
-													number: tag.id,
-													event: undefined,
-													tagArray: tag.name
-												}
-											};
-										}}
-									>
-										{'Menu'}
-									</button>
-								</div>
-							{:else if (tag.name[0] === 'r' && tag.name.length > 1 && tag.name[1].startsWith('ws')) || tag.name[0] === 'relay'}
-								<!--りれー-->
-								<div
-									class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto_auto] gap-1 break-all"
-								>
-									<Relay tagArray={tag.name} />
-									{#if isOwner && $isMulti === MultiMenu.None}
-										<button
-											class="btn p-2 fill-surface-600 dark:fill-surface-300"
-											on:click={() => {
-												handleClick(tag.id, tag.name);
-											}}>{@html EditIcon}</button
-										>
-									{/if}
-									<button
-										class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-										use:popup={popupCombobox}
-										on:click={() => {
-											selectedIndex = {
-												detail: {
-													number: tag.id,
-													event: undefined,
-													tagArray: tag.name
-												}
-											};
-										}}
-									>
-										{'Menu'}
-									</button>
-								</div>
-							{:else if tag.name[0] === 'r'}
-								<div
-									class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto_auto] gap-1 break-all"
-								>
-									<Reference tagArray={tag.name} />
-									{#if isOwner && $isMulti === MultiMenu.None}
-										<button
-											class="btn p-2 fill-surface-600 dark:fill-surface-300"
-											on:click={() => {
-												handleClick(tag.id, tag.name);
-											}}>{@html EditIcon}</button
-										>
-									{/if}
-									<button
-										class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-										use:popup={popupCombobox}
-										on:click={() => {
-											selectedIndex = {
-												detail: {
-													number: tag.id,
-													event: undefined,
-													tagArray: tag.name
-												}
-											};
-										}}
-									>
-										{'Menu'}
-									</button>
-								</div>
-							{:else if tag.name[0] === 't'}
-								<!--はっしゅたぐ-->
-
-								<div
-									class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto_auto] gap-1 break-all"
-								>
-									<Hashtag tagArray={tag.name} />
-									{#if isOwner && $isMulti === MultiMenu.None}
-										<button
-											class="btn p-2 fill-surface-600 dark:fill-surface-300"
-											on:click={() => {
-												handleClick(tag.id, tag.name);
-											}}>{@html EditIcon}</button
-										>
-									{/if}
-									<button
-										class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-										use:popup={popupCombobox}
-										on:click={() => {
-											selectedIndex = {
-												detail: {
-													number: tag.id,
-													event: undefined,
-													tagArray: tag.name
-												}
-											};
-										}}
-									>
-										{'Menu'}
-									</button>
-								</div>
-							{:else if tag.name[0] === 'word'}
-								<!--word-->
-
-								<div
-									class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto_auto] gap-1 break-all"
-								>
-									<Other tagArray={tag.name} />
-									{#if isOwner && $isMulti === MultiMenu.None}
-										<button
-											class="btn p-2 fill-surface-600 dark:fill-surface-300"
-											on:click={() => {
-												handleClick(tag.id, tag.name);
-											}}>{@html EditIcon}</button
-										>
-									{/if}
-									<button
-										class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-										use:popup={popupCombobox}
-										on:click={() => {
-											selectedIndex = {
-												detail: {
-													number: tag.id,
-													event: undefined,
-													tagArray: tag.name
-												}
-											};
-										}}
-									>
-										{'Menu'}
-									</button>
-								</div>
-							{:else}
-								<!--a,e,d,emoji,relay,r,t,word以外-->
-
-								<div
-									class="z-0 card drop-shadow px-1 py-1 my-0.5 grid grid-cols-[1fr_auto_auto] gap-1 break-all"
-								>
-									{JSON.stringify(tag.name)}
-									{#if isOwner && $isMulti === MultiMenu.None}
-										<button
-											class="btn p-2 fill-surface-600 dark:fill-surface-300"
-											on:click={() => {
-												handleClick(tag.id, tag.name);
-											}}>{@html EditIcon}</button
-										>
-									{/if}
-									<button
-										class="fill-white btn btn-sm variant-filled-primary h-fit text-sm p-1"
-										use:popup={popupCombobox}
-										on:click={() => {
-											selectedIndex = {
-												detail: {
-													number: tag.id,
-													event: undefined,
-													tagArray: tag.name
-												}
-											};
-										}}
-									>
-										{'Menu'}
-									</button>
-								</div>
-							{/if}
+							<EventandButtons
+								{id}
+								{tag}
+								{popupCombobox}
+								{pubkey}
+								{isOwner}
+								{filter}
+								bind:selectedIndex
+								{kind}
+								{handleClick}
+								{CheckNote}
+							/>
 						{/if}
 					{/await}
 				</div>
@@ -1085,6 +340,7 @@
 	{/if}
 
 	<div
+		bind:this={popupElement}
 		class="absolute card w-48 shadow-xl py-2 border"
 		data-popup="popupCombobox"
 	>
@@ -1149,6 +405,6 @@
 				njump</ListBoxItem
 			>
 		</ListBox>
-		<div class="arrow bg-surface-100-800-token" />
+		<!-- <div class="arrow bg-surface-100-800-token border" /> -->
 	</div>
 </div>
